Revert "Sbojanic/move yubikey utils v2"
